Use SlowSoftSerial to get lower baud rates
For setup and user interface purposes, the serial interfaces
are still USB and RS232. If the user chooses RS232, all the
lower baud rates in the list are now available. If one of the
lower baud rates is selected, then at the end of setup the
serial mode is changed to RS232_SLOW and the SlowSoftSerial
library object is used instead of the Serial1 hardware UART.

The SlowSoftSerial object itself is instantiated at powerup
and persists forever, but uses only a few bytes of RAM. It
only becomes active when its begin() method is called.

Serial begin and end methods are now abstracted in the same way
that the read, write, available, etc. methods were, so that all
knowledge of serial port multiplexing is in the same part of
the source code.

The user input parsing for baud rates was rewritten in a
different style, because with the larger set of baud rates
there's a case where three characters must be typed before
the entry is unambiguous. The while loop that discards invalid
keystrokes is now in a single place outside the logic.

Updated "Future ideas" list.

Updated block comment about RS232 serial baud rates.

spelling typo patrol

trailing whitespace patrol