Fix handling of serial mode setting in EEPROM (refactored)
Before I had added a third value to `serial` to encode the use of
SlowSoftSerial, but that value couldn't be stored in EEPROM without
breaking backward compatibility and diverging from the semantics in
the UI, so it required some messy fiddling around when reading or
writing EEPROM. Messy enough that I didn't get it right the first
time. Now I have restored the original meaning of `serial` and added
a separate `rs232_mode` variable to keep track of whether we are
using the UART or SlowSoftSerial. This cleans it up, at the expense
of one additional comparison in the serial data path. Worth it.

Also sneaking in here is a workaround for a bug in the Teensy
hardware serial driver. We call Serial1.end() and Serial1.begin()
to change the baud rate, but the driver leaves the receive FIFO
in an indeterminate state in this case. We work around the problem
by calling Serial1.clear() after Serial1.begin().